fix problems

+++ modules/Function_Eyler/src/Function_Eyler_Application.cpp
@@ -26,7 +26,7 @@ std::vector<int> Application::parseFromParam(int argc, char** argv) {
 std::string Application::calculate(std::vector<int> numbs) {
     std::string results;
     Function_Eyler f;
-    for (int i = 0; i < numbs.size(); i++) {
+    for (size_t i = 0; i < numbs.size(); i++) {
         results.append("The Euler function for ");
         results.append(std::to_string(numbs[i]));
         results.append(" = ");
